requirements.txt
Django==5.0.2
djangorestframework==3.14.0
drf-yasg==1.21.7
python-dotenv==1.0.1
Pillow==10.2.0
django-cors-headers==4.3.1 